third_party/WebKit/LayoutTests/platform/android/svg/custom/keyTimes-parsing-error-expected.txt

CONSOLE ERROR: Error: Invalid value for <animate> attribute keyTimes=";;"
CONSOLE ERROR: Error: Invalid value for <animate> attribute keyTimes="0;.25;.5;1;;"
Tests parsing of keyTimes attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
